test(db): separate integration tests from unit tests
DB-dependent tests (TestPing, TestRepositories, TestQueryRow, TestQuery,
TestExec, TestRepositoryErrorSanitization, TestTransaction, TestConnect,
migration and schema validation tests) were mixed into the same package as
pure unit tests with no build tag. Without a live PostgreSQL instance they
would hang for ~20-30s burning through waitForDB retry windows before finally
skipping, causing the test suite to time out.

Changes:
- Move shared DB test infrastructure (getTestConfigs, setupTestDB, waitForDB,
  etc.) into testhelpers_test.go tagged //go:build integration
- Tag migration_test.go and integration_validation_test.go //go:build integration
- Remove the infrastructure helpers and all DB-dependent tests from db_test.go,
  keeping only the three pure unit tests (TestSanitizeDBError,
  TestSanitizeDBErrorPreservesCause, TestConnectionErrorSanitization)

The live-DB tests are already covered by the suite in
test/integration/db/database_integration_test.go which runs under the
integration build tag in CI.

go test ./internal/db/... now completes in ~0.5s with 70 passing tests.
