+++ internal/db/integration_validation_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package db
 
 import (
